requirements.txt
chalice==1.29.0
python-dotenv==1.0.0
marshmallow==3.20.1
pre-commit==3.4.0
pytest==7.4.2
slack_sdk==3.23.0
requests==2.31.0
